+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHN2X34395AG71DS0SDVDACR"
+        "01KHNQ7D72H3J8HVVAZK202CCB"
     )
     r"""The unique identifier of the data source"""